requirements.txt
transformers==5.1.0
accelerate==1.12.0
datasets==4.8.4
optimum==2.1.0
bitsandbytes==0.49.2
peft==0.18.1
trl==1.0.0
torchao==0.15.0
sentencepiece==0.2.1
protobuf==6.33.5
psutil==7.2.2
tyro==1.0.8
wheel==0.46.3
huggingface_hub[hf_xet]==1.5.0
setuptools==81.0.0
dill==0.4.0
httpx==0.28.1
multiprocess==0.70.18
packaging==26.0
pyarrow==23.0.1
pyyaml==6.0.3
requests==2.33.0
tqdm==4.67.3
optuna==4.7.0
mistral-common==1.9.1
tensorboard==2.20.0
